Split AnswersController into VoiceAnswersController and Numerical AnswersController
