+++ app/views/layouts/application.html.erb
@@ -1,81 +1,80 @@
 <!DOCTYPE html>
 <html>
-	<head>
-	  <title>CITYVOICE</title>
-	  <%= stylesheet_link_tag    "application", media: "all", "data-turbolinks-track" => true %>
-	 <link href='http://fonts.googleapis.com/css?family=Varela+Round' rel='stylesheet' type='text/css'>
-	  <%= javascript_include_tag "application", "data-turbolinks-track" => true %>
-	  <%= javascript_include_tag "application", controller_name %>
-	  <%= csrf_meta_tags %>
-		<meta name='viewport' content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no' />
-	  	<link href='http://api.tiles.mapbox.com/mapbox.js/v1.3.1/mapbox.css' rel='stylesheet' />
+  <head>
+    <title>CITYVOICE</title>
+    <%= stylesheet_link_tag    "application", media: "all", "data-turbolinks-track" => true %>
+    <link href='http://fonts.googleapis.com/css?family=Varela+Round' rel='stylesheet' type='text/css'>
+    <%= javascript_include_tag "application", "data-turbolinks-track" => true %>
+    <%= javascript_include_tag "application", controller_name %>
+    <%= csrf_meta_tags %>
+    <meta name='viewport' content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no' />
+    <link href='http://api.tiles.mapbox.com/mapbox.js/v1.3.1/mapbox.css' rel='stylesheet' />
+
     <% if Rails.env == "production" && ENV['GOOGLE_ANALYTICS_ON'] == "true" %>
-		<script type="text/javascript">
-		  var _gaq = _gaq || [];
-		  _gaq.push(['_setAccount', 'UA-43486781-1']);
-		  _gaq.push(['_trackPageview']);
-		  (function() {
-		    var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
-		    ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
-		    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
-		  })();
-      	</script>
+      <script type="text/javascript">
+        var _gaq = _gaq || [];
+        _gaq.push(['_setAccount', 'UA-43486781-1']);
+        _gaq.push(['_trackPageview']);
+        (function() {
+          var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
+          ga.src = ('https:' == document.location.protocol ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js';
+          var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
+        })();
+      </script>
     <% end %>
-        <script>
-          var mapboxMapID = "<%= ENV["MAPBOX_MAP_ID"] %>";
-        </script>
-	</head>
-	<body class="<%= controller_name %> <%= action_name %>" data-no-turbolink>
-		<header class="navbar navbar-fixed-top navbar-inverse">
-                  <div class="navbar-inner" style="background: <%= @content.header_color || "#6DC6AD" %>;  border: none;">
-			    <div class="container">
-	                  <% if [nil, ""].include? @content.short_title %>
-			      	<%= link_to image_tag("cityvoice_icon.png", :border => 0), "/", id: "logo" %>
-	      				<%= link_to "CITYVOICE", "/", id: "logo" %>
-	                  <% else %>
-	                    <%= link_to @content.short_title, "/", id: "logo" %>
-	                  <% end %>
-			      	<nav>
-			        	<ul class="nav" style="margin-top: .35em; margin-left: 2em;">
-				          	<div class="input-append">
-                      			<input id="appendedInputButton" type="text" placeholder="enter address">
-							 <!--<button class="btn" type="button"><i class="icon-search"></i></button>-->
-							</div>
-			        	</ul>
-			      	</nav>
-			      	<nav class="pull-right">
-			        	<ul class="nav" style="margin-top: .6em; margin-left: 2em;">
-			        		<div class="header_links">
-                                <%= link_to "ACTIVITY", most_feedback_path %>
-				          	</div>
-			        		<div class="header_links">
-                                <%= link_to "ALL MESSAGES", voice_messages_path %>
-				          	</div>
-			        	</ul>
-			      	</nav>
-			    </div>
-			</div>
-		</header>
+    <script>
+      var mapboxMapID = "<%= ENV["MAPBOX_MAP_ID"] %>";
+    </script>
+  </head>
+  <body class="<%= controller_name %> <%= action_name %>" data-no-turbolink>
+    <header class="navbar navbar-fixed-top navbar-inverse">
+      <div class="navbar-inner" style="background: <%= @content.header_color || "#6DC6AD" %>;  border: none;">
+       <div class="container">
+          <% if [nil, ""].include? @content.short_title %>
+          <%= link_to image_tag("cityvoice_icon.png", :border => 0), "/", id: "logo" %>
+          <%= link_to "CITYVOICE", "/", id: "logo" %>
+          <% else %>
+          <%= link_to @content.short_title, "/", id: "logo" %>
+          <% end %>
+          <nav>
+            <ul class="nav" style="margin-top: .35em; margin-left: 2em;">
+              <div class="input-append">
+                <input id="appendedInputButton" type="text" placeholder="enter address">
+              </div>
+            </ul>
+          </nav>
+          <nav class="pull-right">
+            <ul class="nav" style="margin-top: .6em; margin-left: 2em;">
+              <div class="header_links">
+                <%= link_to "ACTIVITY", numerical_answers_path %>
+              </div>
+              <div class="header_links">
+                <%= link_to "ALL MESSAGES", voice_answers_path %>
+              </div>
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
 
-		<%= yield %>
+    <%= yield %>
 
-        <% unless @content.feedback_form_url.empty? %>
-          <div id="feedback">
-            <a href="<%= @content.feedback_form_url %>" target="_blank"><img src="/assets/feedback_button.png" /></a>
-          </div>
-        <% end %>
+    <% unless @content.feedback_form_url.empty? %>
+      <div id="feedback">
+        <a href="<%= @content.feedback_form_url %>" target="_blank"><img src="/assets/feedback_button.png" /></a>
+      </div>
+    <% end %>
 
-		<footer class="footer">
-		  <nav class="pull right" style="padding-right: 1.5em;">
-		    <ul>
-		      <li><%= link_to "Contact", page_path('contact') %></li>
-		      <li><%= link_to "About", page_path('about') %></li>
-		    </ul>
-		  </nav>
-		  <div class="footerLogo"><a href="http://www.cityvoiceapp.com/"><img src="/assets/CityVoice_footer.png" /></a></div>
-		  <div class="footerLogo"><a href="http://codeforamerica.org/"><img src="/assets/cfa_logo_black.png" /></a></div>
-                  <div class="footerLogo"><a href="http://southbendin.gov/"><img src="/assets/southbendlogo.png" /><strong> City of South Bend</strong></a></div>
-		</footer>
-	</body>
+    <footer class="footer">
+      <nav class="pull right" style="padding-right: 1.5em;">
+        <ul>
+          <li><%= link_to "Contact", page_path('contact') %></li>
+          <li><%= link_to "About", page_path('about') %></li>
+        </ul>
+      </nav>
+      <div class="footerLogo"><a href="http://www.cityvoiceapp.com/"><img src="/assets/CityVoice_footer.png" /></a></div>
+      <div class="footerLogo"><a href="http://codeforamerica.org/"><img src="/assets/cfa_logo_black.png" /></a></div>
+      <div class="footerLogo"><a href="http://southbendin.gov/"><img src="/assets/southbendlogo.png" /><strong> City of South Bend</strong></a></div>
+    </footer>
+  </body>
 </html>
-
